[REFACTORING] Relocate LDAP retrying logic from server-utils into data-ldap
