+++ server/data/data-ldap/src/main/java/org/apache/james/user/ldap/retry/ExceptionRetryHandler.java
@@ -18,11 +18,11 @@
  *
  */
 
-package org.apache.james.util.retry;
+package org.apache.james.user.ldap.retry;
 
-import org.apache.james.util.retry.api.ExceptionRetryingProxy;
-import org.apache.james.util.retry.api.RetryHandler;
-import org.apache.james.util.retry.api.RetrySchedule;
+import org.apache.james.user.ldap.retry.api.ExceptionRetryingProxy;
+import org.apache.james.user.ldap.retry.api.RetryHandler;
+import org.apache.james.user.ldap.retry.api.RetrySchedule;
 
 /**
  * Abstract class <code>ExceptionRetryHandler</code> retries the behaviour defined in abstract method 
@@ -36,7 +36,7 @@ import org.apache.james.util.retry.api.RetrySchedule;
  * The method stubs required to perform the proxy call via <code>operation()</code> may be generated by many means,
  * including explicit code, a <code>DynamicProxy</code> and compile time aspect injection.
  *
- * @see org.apache.james.util.retry.naming.RetryingContext
+ * @see org.apache.james.user.ldap.retry.naming.RetryingContext
  */
 public abstract class ExceptionRetryHandler implements RetryHandler {
